requirements.txt
psutil>=5.9.0
requests>=2.28.0
python-dotenv>=1.0.0
pytest>=7.4.0
rich>=13.0.0
